+++ vcfpy/__init__.py
@@ -24,4 +24,4 @@ from .reader import Reader
 
 from .writer import Writer
 
-__version__ = '0.7.0'
+__version__ = '0.8.0'