+++ vcfpy/header.py
@@ -301,8 +301,9 @@ class Header:
         :param mapping: ``OrderedDict`` with mapping to add.  It is
             recommended to use ``OrderedDict`` over ``dict`` as this makes
             the result reproducible
+        :return: ``False`` on conflicting line and ``True`` otherwise
         """
-        self.add_line(FilterHeaderLine.from_mapping(mapping))
+        return self.add_line(FilterHeaderLine.from_mapping(mapping))
 
     def add_contig_line(self, mapping):
         """Add "contig" header line constructed from the given mapping
@@ -310,8 +311,9 @@ class Header:
         :param mapping: ``OrderedDict`` with mapping to add.  It is
             recommended to use ``OrderedDict`` over ``dict`` as this makes
             the result reproducible
+        :return: ``False`` on conflicting line and ``True`` otherwise
         """
-        self.add_line(ContigHeaderLine.from_mapping(mapping))
+        return self.add_line(ContigHeaderLine.from_mapping(mapping))
 
     def add_info_line(self, mapping):
         """Add INFO header line constructed from the given mapping
@@ -319,8 +321,9 @@ class Header:
         :param mapping: ``OrderedDict`` with mapping to add.  It is
             recommended to use ``OrderedDict`` over ``dict`` as this makes
             the result reproducible
+        :return: ``False`` on conflicting line and ``True`` otherwise
         """
-        self.add_line(InfoHeaderLine.from_mapping(mapping))
+        return self.add_line(InfoHeaderLine.from_mapping(mapping))
 
     def add_format_line(self, mapping):
         """Add FORMAT header line constructed from the given mapping
@@ -328,8 +331,9 @@ class Header:
         :param mapping: ``OrderedDict`` with mapping to add.  It is
             recommended to use ``OrderedDict`` over ``dict`` as this makes
             the result reproducible
+        :return: ``False`` on conflicting line and ``True`` otherwise
         """
-        self.add_line(FormatHeaderLine.from_mapping(mapping))
+        return self.add_line(FormatHeaderLine.from_mapping(mapping))
 
     def format_ids(self):
         """Return list of all format IDs"""
@@ -350,20 +354,40 @@ class Header:
         else:
             return []
 
+    def has_header_line(self, key, id_):
+        """Return whether there is a header line with the given ID of the
+        type given by ``key``
+
+        :param key: The VCF header key/line type.
+        :param id_: The ID value to compare fore
+
+        :return: ``True`` if there is a header line starting with ``##${key}=``
+            in the VCF file having the mapping entry ``ID`` set to ``id_``.
+        """
+        if key not in self._indices:
+            return False
+        else:
+            return id_ in self._indices[key]
+
     def add_line(self, header_line):
-        """Add header line, updating any necessary support indices"""
+        """Add header line, updating any necessary support indices
+
+        :return: ``False`` on conflicting line and ``True`` otherwise
+        """
         self.lines.append(header_line)
         self._indices.setdefault(header_line.key, OrderedDict())
         if not hasattr(header_line, 'mapping'):
-            return  # no registration required
-        if header_line.mapping['ID'] in self._indices[header_line.key]:
+            return False  # no registration required
+        if self.has_header_line(header_line.key, header_line.mapping['ID']):
             self.warning_helper.warn_once(
                 ('Detected duplicate header line with type {} and ID {}. '
                  'Ignoring this and subsequent one').format(
                      header_line.key, header_line.mapping['ID']))
+            return False
         else:
             self._indices[header_line.key][
                 header_line.mapping['ID']] = header_line
+            return True
 
     def get_info_field_info(self, key):
         """Return :py:class:`FieldInfo` for the given INFO field"""