+++ HISTORY.rst
@@ -2,6 +2,13 @@
 History
 =======
 
+0.8.0 (2016-10-31)
+------------------
+
+* Adding ``Header.has_header_line`` for querying existence of header line
+* ``Header.add_*_line`` return a ``bool`` no indicating any conflicts
+* Construction of Writer uses samples within header and no extra parameter (breaks API)
+
 0.7.0 (2016-09-25)
 ------------------